+++ src/components/GameHUD/Score.tsx
@@ -4,12 +4,16 @@ import "./Score.css"
 
 const Score = ()=>{
 
+    const gamePhase = useGameStore(state => state.gamePhase);
     const gameMode = useGameStore(state => state.gameMode);
+
     const playerScore = useScoreStore(state => state.playerScore);
     const playerCombo = useScoreStore(state => state.playerCombo);
     const comScore = useScoreStore(state => state.comScore);
     const comCombo = useScoreStore(state => state.comCombo);
 
+    if(gamePhase === "welcome") return null;
+    
     return(
         <div className="Score">
             <div className="score__row">